+++ src/main/java/courgette/integration/extentreports/ExtentReportsBuilder.java
@@ -8,7 +8,11 @@ import com.aventstack.extentreports.reporter.ExtentSparkReporter;
 import courgette.runtime.CourgetteException;
 import courgette.runtime.report.model.*;
 
+import java.time.Instant;
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 
 public class ExtentReportsBuilder {
@@ -34,6 +38,7 @@ public class ExtentReportsBuilder {
         }
 
         final ExtentReports extentReports = new ExtentReports();
+        extentReports.setReportUsesManualConfiguration(true);
         extentReports.attachReporter(extentSparkReporter);
 
         getDistinctFeatureUris().forEach(featureUri -> {
@@ -51,57 +56,74 @@ public class ExtentReportsBuilder {
         final ExtentTest featureNode = extentReports.createTest(features.get(0).getName());
 
         features.forEach(feature -> {
-            if (feature.getScenarios().size() == 1) {
-                Scenario scenario = feature.getScenarios().get(0);
-                addScenario(featureNode, scenario.getName(), scenario.getBefore(), scenario.getAfter(), scenario.getSteps(), scenario.getTags());
-            } else {
-                for (Scenario scenario : feature.getScenarios()) {
-                    if (scenario.getKeyword().startsWith("Scenario")) {
-                        addScenario(featureNode, scenario.getName(), scenario.getBefore(), scenario.getAfter(), scenario.getSteps(), scenario.getTags());
-                    }
+            if (feature.getScenarios().size() > 1) {
+                addFeatureStartAndEndTime(featureNode, feature);
+            }
+
+            for (Scenario scenario : feature.getScenarios()) {
+                if (scenario.getKeyword().startsWith("Scenario")) {
+                    addScenario(featureNode, scenario);
+                    addScenarioStartAndEndTime(featureNode, scenario);
                 }
             }
         });
     }
 
-    private void addScenario(ExtentTest featureNode, String scenarioName, List<Hook> before, List<Hook> after, List<Step> steps, List<Tag> tags) {
-        final ExtentTest scenarioNode = createGherkinNode(featureNode, "Scenario", scenarioName, false);
-        addBeforeOrAfterDetails(scenarioNode, before);
-        addSteps(scenarioNode, steps);
-        addBeforeOrAfterDetails(scenarioNode, after);
-        assignCategoryToScenario(scenarioNode, tags);
+    private void addScenario(ExtentTest featureNode, Scenario scenario) {
+        final ExtentTest scenarioNode = createGherkinNode(featureNode, "Scenario", scenario.getName(), false);
+        addBeforeOrAfterDetails(scenarioNode, scenario.getBefore());
+        addSteps(scenarioNode, scenario);
+        addBeforeOrAfterDetails(scenarioNode, scenario.getAfter());
+        assignCategoryToScenario(scenarioNode, scenario.getTags());
     }
 
     private void assignCategoryToScenario(ExtentTest scenarioNode, List<Tag> tags) {
         tags.forEach(tag -> scenarioNode.assignCategory(tag.getName()));
     }
 
-    private void addSteps(ExtentTest scenarioNode, List<Step> steps) {
+    private void addSteps(ExtentTest scenarioNode, Scenario scenario) {
+        Date startTime = getStartTime(scenario.getStartTimestamp());
+        List<Step> steps = scenario.getSteps();
+
         steps.forEach(step -> {
             addBeforeOrAfterDetails(scenarioNode, step.getBefore());
-            ExtentTest stepNode = createGherkinNode(scenarioNode, step.getKeyword().trim(), step.getName(), true);
+            ExtentTest stepNode = createStepNode(scenarioNode, step, startTime);
             addAdditionalStepDetails(stepNode, step);
             setStepResult(stepNode, step.passed(isStrict));
             addBeforeOrAfterDetails(scenarioNode, step.getAfter());
         });
     }
 
+    private ExtentTest createStepNode(ExtentTest scenarioNode, Step step, Date startTime) {
+        ExtentTest stepNode = createGherkinNode(scenarioNode, step.getKeyword().trim(), step.getName(), true);
+        stepNode.getModel().setStartTime(startTime);
+        return stepNode;
+    }
+
     private void addBeforeOrAfterDetails(ExtentTest scenarioNode, List<Hook> hooks) {
         hooks.forEach(hook -> {
-            addOutputs(scenarioNode, hook.getOutput());
-            addError(scenarioNode, hook.getResult().getErrorMessage());
-            addImageEmbeddings(scenarioNode, hook.getEmbeddings());
+            String error = hook.getResult().getErrorMessage();
+            List<String> output = hook.getOutput();
+            List<Embedding> embeddings = hook.getEmbeddings();
+
+            addOutputs(scenarioNode, output);
+            addError(scenarioNode, error);
+            addImageEmbeddings(scenarioNode, embeddings);
         });
     }
 
     private void addAdditionalStepDetails(ExtentTest stepNode, Step step) {
-        addOutputs(stepNode, step.getOutput());
-        addError(stepNode, step.getResult().getErrorMessage());
-        addImageEmbeddings(stepNode, step.getEmbeddings());
+        String error = step.getResult().getErrorMessage();
+        List<String> output = step.getOutput();
+        List<Embedding> embeddings = step.getEmbeddings();
+
+        addOutputs(stepNode, output);
+        addError(stepNode, error);
+        addImageEmbeddings(stepNode, embeddings);
     }
 
-    private void addOutputs(ExtentTest scenarioNode, List<String> outputs) {
-        outputs.forEach(output -> log(scenarioNode, output));
+    private void addOutputs(ExtentTest node, List<String> outputs) {
+        outputs.forEach(output -> log(node, output));
     }
 
     private void addImageEmbeddings(ExtentTest node, List<Embedding> embeddings) {
@@ -112,8 +134,8 @@ public class ExtentReportsBuilder {
         });
     }
 
-    private void addError(ExtentTest scenarioNode, String error) {
-        log(scenarioNode, error);
+    private void addError(ExtentTest node, String error) {
+        log(node, error);
     }
 
     private void log(ExtentTest node, String message) {
@@ -136,4 +158,48 @@ public class ExtentReportsBuilder {
             extentTest.fail("Step failed");
         }
     }
+
+    private void addFeatureStartAndEndTime(ExtentTest featureNode, Feature feature) {
+        Date featureStartTime = getEarliestStartTime(feature.getScenarios());
+        Date featureEndTime = getEndTime(featureStartTime.getTime(), feature.getScenarios());
+        addStartAndEndTime(featureNode, featureStartTime, featureEndTime);
+    }
+
+    private void addScenarioStartAndEndTime(ExtentTest scenarioNode, Scenario scenario) {
+        List<Scenario> scenarios = new ArrayList<>();
+        scenarios.add(scenario);
+        Date scenarioStartTime = getStartTime(scenario.getStartTimestamp());
+        Date scenarioEndTime = getEndTime(scenarioStartTime.getTime(), scenarios);
+        addStartAndEndTime(scenarioNode, scenarioStartTime, scenarioEndTime);
+    }
+
+    private void addStartAndEndTime(ExtentTest node, Date startTime, Date endTime) {
+        node.getModel().setStartTime(startTime);
+        node.getModel().setEndTime(endTime);
+    }
+
+    private Date getEarliestStartTime(List<Scenario> scenarios) {
+        List<Long> times = new ArrayList<>();
+        scenarios.stream().map(Scenario::getStartTimestamp).filter(time -> time.length() > 0).forEach(time -> times.add(Date.from(Instant.parse(time)).getTime()));
+        return new Date(times.stream().reduce((start, end) -> start).get());
+    }
+
+    private Date getStartTime(String timestamp) {
+        return Date.from(Instant.parse(timestamp));
+    }
+
+    private Date getEndTime(long startTime, List<Scenario> scenarios) {
+        long endTime = 0;
+
+        for (Scenario scenario : scenarios) {
+            endTime = endTime + calculateDuration.apply(scenario.getBefore().stream().map(Hook::getResult).collect(Collectors.toList()));
+            endTime = endTime + calculateDuration.apply(scenario.getAfter().stream().map(Hook::getResult).collect(Collectors.toList()));
+            endTime = endTime + calculateDuration.apply(scenario.getSteps().stream().map(Step::getResult).collect(Collectors.toList()));
+            endTime = endTime + calculateDuration.apply(scenario.getSteps().stream().flatMap(s -> s.getBefore().stream()).map(Hook::getResult).collect(Collectors.toList()));
+            endTime = endTime + calculateDuration.apply(scenario.getSteps().stream().flatMap(s -> s.getAfter().stream()).map(Hook::getResult).collect(Collectors.toList()));
+        }
+        return new Date(startTime + endTime);
+    }
+
+    private Function<List<Result>, Long> calculateDuration = (source) -> source.stream().mapToLong(Result::getDuration).sum();
 }